third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/serviceworker/fetch-response-taint-expected.txt

CONSOLE WARNING: line 14: Subresource requests whose URLs contain embedded credentials (e.g. `https://user:pass@host/`) are deprecated, and will be blocked in M59, around June 2017. See https://www.chromestatus.com/feature/5669008342777856 for more details.
CONSOLE ERROR: line 18: Fetch API cannot load http://localhost:8000/serviceworker/resources/fetch-access-control.php?. Request mode is "same-origin" but the URL's origin is not same as the request origin http://127.0.0.1:8000.
CONSOLE ERROR: line 18: Fetch API cannot load http://localhost:8000/serviceworker/resources/fetch-access-control.php?. Request mode is "same-origin" but the URL's origin is not same as the request origin http://127.0.0.1:8000.
CONSOLE ERROR: line 18: Fetch API cannot load http://localhost:8000/serviceworker/resources/fetch-access-control.php?. Request mode is "same-origin" but the URL's origin is not same as the request origin http://127.0.0.1:8000.
CONSOLE ERROR: line 18: Fetch API cannot load http://localhost:8000/?url=http%3A%2F%2F127.0.0.1%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3F&mode=same-origin&credentials=omit&. Request mode is "same-origin" but the URL's origin is not same as the request origin http://127.0.0.1:8000.
CONSOLE ERROR: line 18: Fetch API cannot load http://localhost:8000/?url=http%3A%2F%2F127.0.0.1%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3F&mode=same-origin&credentials=omit&. Request mode is "same-origin" but the URL's origin is not same as the request origin http://127.0.0.1:8000.
CONSOLE ERROR: line 18: Fetch API cannot load http://localhost:8000/?url=http%3A%2F%2F127.0.0.1%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3F&mode=same-origin&credentials=omit&. Request mode is "same-origin" but the URL's origin is not same as the request origin http://127.0.0.1:8000.
CONSOLE ERROR: line 18: Fetch API cannot load http://localhost:8000/?url=http%3A%2F%2F127.0.0.1%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3F&mode=same-origin&credentials=same-origin&. Request mode is "same-origin" but the URL's origin is not same as the request origin http://127.0.0.1:8000.
CONSOLE ERROR: line 18: Fetch API cannot load http://localhost:8000/?url=http%3A%2F%2F127.0.0.1%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3F&mode=same-origin&credentials=same-origin&. Request mode is "same-origin" but the URL's origin is not same as the request origin http://127.0.0.1:8000.
CONSOLE ERROR: line 18: Fetch API cannot load http://localhost:8000/?url=http%3A%2F%2F127.0.0.1%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3F&mode=same-origin&credentials=same-origin&. Request mode is "same-origin" but the URL's origin is not same as the request origin http://127.0.0.1:8000.
CONSOLE ERROR: line 18: Fetch API cannot load http://localhost:8000/?url=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3F&mode=same-origin&credentials=omit&. Request mode is "same-origin" but the URL's origin is not same as the request origin http://127.0.0.1:8000.
CONSOLE ERROR: line 18: Fetch API cannot load http://localhost:8000/?url=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3F&mode=same-origin&credentials=omit&. Request mode is "same-origin" but the URL's origin is not same as the request origin http://127.0.0.1:8000.
CONSOLE ERROR: line 18: Fetch API cannot load http://localhost:8000/?url=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3F&mode=same-origin&credentials=omit&. Request mode is "same-origin" but the URL's origin is not same as the request origin http://127.0.0.1:8000.
CONSOLE ERROR: line 18: Fetch API cannot load http://localhost:8000/?url=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3F&mode=no-cors&credentials=omit&. Request mode is "same-origin" but the URL's origin is not same as the request origin http://127.0.0.1:8000.
CONSOLE ERROR: line 18: Fetch API cannot load http://localhost:8000/?url=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3F&mode=no-cors&credentials=omit&. Request mode is "same-origin" but the URL's origin is not same as the request origin http://127.0.0.1:8000.
CONSOLE ERROR: line 18: Fetch API cannot load http://localhost:8000/?url=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3F&mode=no-cors&credentials=omit&. Request mode is "same-origin" but the URL's origin is not same as the request origin http://127.0.0.1:8000.
CONSOLE ERROR: line 18: Fetch API cannot load http://localhost:8000/?url=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*&mode=cors&credentials=omit&. Request mode is "same-origin" but the URL's origin is not same as the request origin http://127.0.0.1:8000.
CONSOLE ERROR: line 18: Fetch API cannot load http://localhost:8000/?url=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*&mode=cors&credentials=omit&. Request mode is "same-origin" but the URL's origin is not same as the request origin http://127.0.0.1:8000.
CONSOLE ERROR: line 18: Fetch API cannot load http://localhost:8000/?url=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*&mode=cors&credentials=omit&. Request mode is "same-origin" but the URL's origin is not same as the request origin http://127.0.0.1:8000.
CONSOLE ERROR: line 18: Fetch API cannot load http://localhost:8000/?url=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3Dhttp%3A%2F%2F127.0.0.1%3A8000%26ACACredentials%3Dtrue&mode=cors&credentials=include&. Request mode is "same-origin" but the URL's origin is not same as the request origin http://127.0.0.1:8000.
CONSOLE ERROR: line 18: Fetch API cannot load http://localhost:8000/?url=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3Dhttp%3A%2F%2F127.0.0.1%3A8000%26ACACredentials%3Dtrue&mode=cors&credentials=include&. Request mode is "same-origin" but the URL's origin is not same as the request origin http://127.0.0.1:8000.
CONSOLE ERROR: line 18: Fetch API cannot load http://localhost:8000/?url=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3Dhttp%3A%2F%2F127.0.0.1%3A8000%26ACACredentials%3Dtrue&mode=cors&credentials=include&. Request mode is "same-origin" but the URL's origin is not same as the request origin http://127.0.0.1:8000.
This is a testharness.js-based test.
PASS Verify the tainting of responses fetched via SW 
Harness: the test ran to completion.
